# Conflicts:
#	src/sardana/__init__.py
#	src/sardana/macroserver/msexception.py
#	src/sardana/macroserver/recorders/storage.py
#	src/sardana/macroserver/scan/gscan.py
#	src/sardana/taurus/core/tango/sardana/pool.py
#	src/sardana/taurus/core/tango/sardana/sardana.py

+++ src/sardana/pool/poolcontrollers/IoverI0.py
@@ -44,7 +44,7 @@ class IoverI0(PseudoCounterController):
     def Calc(self, axis, counter_values):
         i, i0 = counter_values
         try:
-            i = float(i / i0)
+            i = i / i0
         except ZeroDivisionError:
             pass
         return i
